目前可以将给从机写入的数据再读出来进行比较的功能。并实现了从机升级功能。下一步需要优化屏幕显示升级状态的功能。

+++ App/Uart_communication.h
@@ -26,7 +26,7 @@
 #define     S_W_ADD           0x02
 #define     S_W_NUM           0x03
 
-#define     MCU_START_FLASH   (0x08000000)
+#define     MCU_START_FLASH   ((uint32_t)0x08000000)
